refactor: move logic to service
